requirements.txt
nicegui
bcrypt
pyjwt
psycopg2-binary
redis
fastapi
uvicorn
